[FIXED] - removed unnecassary code, functions work when compiling, new function names for better understanding
