+++ headers/tokener.h
@@ -6,10 +6,9 @@
 /*   By: iatilla- <iatilla-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: Invalid date        by                   #+#    #+#             */
-/*   Updated: 2025/04/25 00:27:18 by iatilla-         ###   ########.fr       */
+/*   Updated: 2025/04/25 01:12:33 by iatilla-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-
 
 #ifndef TOKENER_H
 # define TOKENER_H
@@ -48,6 +47,7 @@ int				has_env_vars(char *str);
 int				handle_operator_segment(t_parse_params *params, int i);
 void			process_segment(t_parse_params *params);
 void			handle_segment(t_parse_params *params, int i);
+int	process_no_quote_ops(char *input, t_token **tokens, int position);
 
 /* --- Parse State Init move to Parser header with First Occasion --- */
 void			init_parse_state(t_parse_state *state, t_token **tokens);
@@ -60,7 +60,7 @@ int				handle_without_quotes(char *input, t_token **tokens, int i);
 char			*handle_escapes(char *input);
 int				handle_whitespace(char *input, t_parse_state *state);
 int				handle_backslash(char *input, t_parse_state *state);
-
+int				handle_quotes(char *input, t_parse_state *state);
 /* --- Token Type Logic --- */
 
 int				is_builtin_command(char *token);
@@ -89,7 +89,6 @@ int				handle_quotes_tokenize(char *input, int i, int *in_quote,
 int				handle_escape(char *input, int i);
 int				is_operator(char *input, int i);
 
-int				process_operator_token(char *input, t_token **tokens, int j);
 int				handle_quoted_string(char *str, int i, t_token_type quote_type,
 					int *error);
