+++ sources/execution/env_checker.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   env_checker.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: schiper <schiper@student.42.fr>            +#+  +:+       +#+        */
+/*   By: iatilla- <iatilla-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/23 22:54:23 by iatilla-          #+#    #+#             */
-/*   Updated: 2025/04/24 18:35:37 by schiper          ###   ########.fr       */
+/*   Updated: 2025/04/25 00:38:03 by iatilla-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
