+++ sources/memory_management/graph_free.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   graph_free.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: schiper <schiper@student.42.fr>            +#+  +:+       +#+        */
+/*   By: iatilla- <iatilla-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/22 13:56:46 by schiper           #+#    #+#             */
-/*   Updated: 2025/04/22 14:17:10 by schiper          ###   ########.fr       */
+/*   Updated: 2025/04/25 01:14:20 by iatilla-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
