+++ sources/tokener/utils/token_interface/token_helper1.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   token_helper1.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: schiper <schiper@student.42.fr>            +#+  +:+       +#+        */
+/*   By: iatilla- <iatilla-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/21 21:30:23 by schiper           #+#    #+#             */
-/*   Updated: 2025/04/21 22:00:05 by schiper          ###   ########.fr       */
+/*   Updated: 2025/04/25 01:14:37 by iatilla-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
